--- packages/core/test/lib/envelope.test.ts
@@ -1,79 +1,0 @@
-import type { DsnComponents, DynamicSamplingContext, Event } from '@sentry/types';
-
-import { createEventEnvelope } from '../../src/envelope';
-
-const testDsn: DsnComponents = { protocol: 'https', projectId: 'abc', host: 'testry.io', publicKey: 'pubKey123' };
-
-describe('createEventEnvelope', () => {
-  describe('trace header', () => {
-    const testTable: Array<[string, Event, DynamicSamplingContext]> = [
-      [
-        'adds minimal baggage items',
-        {
-          type: 'transaction',
-          sdkProcessingMetadata: {
-            dynamicSamplingContext: { trace_id: '1234', public_key: 'pubKey123' },
-          },
-        },
-        { trace_id: '1234', public_key: 'pubKey123' },
-      ],
-      [
-        'adds multiple baggage items',
-        {
-          type: 'transaction',
-          sdkProcessingMetadata: {
-            dynamicSamplingContext: {
-              environment: 'prod',
-              release: '1.0.0',
-              public_key: 'pubKey123',
-              trace_id: '1234',
-            },
-          },
-        },
-        { release: '1.0.0', environment: 'prod', trace_id: '1234', public_key: 'pubKey123' },
-      ],
-      [
-        'adds all baggage items',
-        {
-          type: 'transaction',
-          sdkProcessingMetadata: {
-            dynamicSamplingContext: {
-              environment: 'prod',
-              release: '1.0.0',
-              transaction: 'TX',
-              user_segment: 'segmentA',
-              sample_rate: '0.95',
-              public_key: 'pubKey123',
-              trace_id: '1234',
-            },
-          },
-        },
-        {
-          environment: 'prod',
-          release: '1.0.0',
-          transaction: 'TX',
-          user_segment: 'segmentA',
-          sample_rate: '0.95',
-          public_key: 'pubKey123',
-          trace_id: '1234',
-        },
-      ],
-      [
-        'with error event',
-        {
-          sdkProcessingMetadata: {
-            dynamicSamplingContext: { trace_id: '1234', public_key: 'pubKey123' },
-          },
-        },
-        { trace_id: '1234', public_key: 'pubKey123' },
-      ],
-    ];
-    it.each(testTable)('%s', (_: string, event, trace) => {
-      const envelopeHeaders = createEventEnvelope(event, testDsn)[0];
-
-      expect(envelopeHeaders).toBeDefined();
-      expect(envelopeHeaders.trace).toBeDefined();
-      expect(envelopeHeaders.trace).toEqual(trace);
-    });
-  });
-});
